backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
azure-ai-textanalytics>=5.3.0
python-docx>=1.0.0
PyPDF2>=3.0.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
python-Levenshtein>=0.21.0 